CivicSync-v2-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.40.0
pydantic==2.5.3

pdfplumber==0.11.4
textstat==0.7.3

rank-bm25==0.2.2

anthropic==0.49.0
voyageai==0.3.2

python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.4

pytest==7.4.3
pytest-asyncio==0.21.1
